safari-600.7-branch/LayoutTests/platform/mac/accessibility/table-with-zebra-rows-expected.txt

'$scr'	scripting additions folder	Property of class: domain	System Events : Disk-Folder-File Suite	1
'%doc'	music folder	Property of class: user domain object	System Events : Disk-Folder-File Suite	2
'µdoc'	music folder	Enumeration	Standard Additions : File Commands	3
'ƒhlp'	help	Enumeration	Standard Additions : File Commands	4
'ƒhlp'	help folder	Enumeration	Standard Additions : File Commands	5
'ƒlib'	shared libraries	Enumeration	Standard Additions : File Commands	6
This tests that tables with zebra striped rows are exposed as AXTables.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.role is 'AXRole: AXTable'
PASS successfullyParsed is true

TEST COMPLETE
